NLM/requirements.txt
# NLM Distillation Framework Requirements
# Python 3.10+ required

# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=4.0.0
accelerate>=0.22.0
peft>=0.4.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Data processing
numpy>=1.24.0

# Monitoring (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Inference server
flask>=2.3.0
gunicorn>=21.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# AWS/SageMaker (optional)
sagemaker>=2.150.0
boto3>=1.26.0
botocore>=1.29.0

